move product to own folder
